Add permissions-based visibility and enforcement for file actions in File Manager. (#20)

+++ src/Concerns/HandlesBulkOperations.php
@@ -4,6 +4,7 @@ namespace MmesDesign\FilamentFileManager\Concerns;
 
 use Filament\Actions\Action;
 use Filament\Notifications\Notification;
+use MmesDesign\FilamentFileManager\FileManagerPlugin;
 use MmesDesign\FilamentFileManager\Forms\Components\FolderTreePicker;
 
 trait HandlesBulkOperations
@@ -14,10 +15,13 @@ trait HandlesBulkOperations
             ->label(__('filament-file-manager::file-manager.actions.delete_selected'))
             ->icon('heroicon-o-trash')
             ->color('danger')
+            ->visible(FileManagerPlugin::get()->canUserDelete())
             ->requiresConfirmation()
             ->modalHeading(__('filament-file-manager::file-manager.modals.confirm_deletion'))
             ->modalDescription(fn (): string => __('filament-file-manager::file-manager.modals.bulk_deletion_warning', ['count' => count($this->selectedItems)]))
             ->action(function (): void {
+                abort_unless(FileManagerPlugin::get()->canUserDelete(), 403, __('filament-file-manager::file-manager.messages.permission_denied'));
+
                 $count = $this->fileManagerService->deleteBulk($this->currentDisk, $this->selectedItems);
                 $this->selectedItems = [];
 
@@ -34,6 +38,7 @@ trait HandlesBulkOperations
             ->label(__('filament-file-manager::file-manager.actions.move_selected'))
             ->icon('heroicon-o-arrow-right')
             ->color('gray')
+            ->visible(FileManagerPlugin::get()->canUserMove())
             ->schema([
                 FolderTreePicker::make('destination')
                     ->label(__('filament-file-manager::file-manager.labels.destination_folder'))
@@ -41,6 +46,8 @@ trait HandlesBulkOperations
                     ->default(''),
             ])
             ->action(function (array $data): void {
+                abort_unless(FileManagerPlugin::get()->canUserMove(), 403, __('filament-file-manager::file-manager.messages.permission_denied'));
+
                 $destination = $data['destination'] ?? '';
                 $count = $this->fileManagerService->moveBulk($this->currentDisk, $this->selectedItems, $destination);
                 $this->selectedItems = [];
